Move copypasted private DF code to `datafusion`
